+++ src/middleware/services.ts
@@ -2,7 +2,7 @@ import { NextFunction, Request, Response } from 'express';
 
 import { Locale } from '../enums/locale';
 import { DatasetService } from '../services/dataset';
-import { getFileService } from '../utils/get-storage';
+import { getFileService } from '../utils/get-file-service';
 
 // initialise any request-scoped services required by the app and store them on the request object for later use
 // see @types/express/index.d.ts for details
